Move native stuff in light client to services
